pyproject.toml
[project]
name = "ideas-bot"
version = "0.1.0"
description = "Bot for create and share ideas with your friend"
authors = [
    {name = "Tarasov Artyom",email = "almtara550@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiogram (>=3.22.0,<4.0.0)",
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.0",
    "alembic>=1.13.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
start = "bot.main:main"
